requirements.txt
asttokens==2.4.1
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.45.1
ipython==8.18.0
ipywidgets==8.1.1
jedi==0.19.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
llvmlite==0.41.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
numba==0.58.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
patsy==0.5.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.0
sympy==1.12
tqdm==4.66.1
traitlets==5.13.0
tzdata==2023.3
wcwidth==0.2.12
widgetsnbextension==4.0.9
joblib==1.3.2
openpyxl==3.1.2
